feat: implement canvas redrawing when changing r
