+++ lab4/frontend/src/reducers/rReducer.js
@@ -1,0 +1,14 @@
+const initialState = {
+    r: null,
+};
+
+function rReducer(state = initialState, action) {
+    switch (action.type) {
+        case 'SET_R':
+            return {...state, r: action.payload};
+        default:
+            return state;
+    }
+}
+
+export default rReducer;
